+++ src/vec3/ft_len_vec3.c
@@ -6,16 +6,16 @@
 /*   By: wkorande <willehard@gmail.com>             +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/01/16 14:04:52 by wkorande          #+#    #+#             */
-/*   Updated: 2020/11/13 23:02:10 by wkorande         ###   ########.fr       */
+/*   Updated: 2020/11/15 11:47:56 by wkorande         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "vec3.h"
 #include <math.h>
 
-double	ft_len_vec3(t_vec3 v)
+float	ft_len_vec3(t_vec3 v)
 {
-	double l;
+	float l;
 
 	if (v.x == 0.0f && v.y == 0.0f && v.z == 0.0f)
 		return (0.0f);
